requirements.txt
# 核心依赖
numpy>=1.22.0
scipy>=1.8.0
pandas>=1.4.0
scikit-learn>=1.0.0

# 解释算法库
shap>=0.41.0
lime>=0.2.0.1
dice-ml>=0.8
captum>=0.5.0  # PyTorch解释

# 深度学习框架
torch>=1.12.0
tensorflow>=2.8.0
keras>=2.8.0

# 数据处理与可视化
matplotlib>=3.5.0
seaborn>=0.11.2
Pillow>=9.0.0
plotly>=5.8.0

# CLI与工具
argcomplete>=2.0.0
tqdm>=4.62.0
PyYAML>=6.0
click>=8.0.0
rich>=12.0.0

# 报告生成
Jinja2>=3.0.0
pdfkit>=1.0.0
weasyprint>=54.0

# 测试与文档
pytest>=7.0.0
coverage>=6.0.0
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# 其他工具
ipython>=8.0.0
jupyter>=1.0.0